Backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
boto3
google-genai
pypdf
tenacity
python-dotenv
python-multipart
